requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-bootstrap3==9.1.0
django-social-share==1.1.2
django-tinymce4-lite==1.7.0
docopt==0.6.2
idna==2.6
jsmin==2.2.2
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
yarg==0.1.9
